Add Remark Feature
Add ability to create an remark on an existing user.

Remark can only be added on an existing user.
It can be added and removed.

It is a string, it does not have any validation checks as the remark can
 be any string.

Also add test for the new remark command.

Updated older test to ensure remark command works on previous tests.